backend/requirements.txt
functions-framework==3.*
Pillow==10.*
google-generativeai==0.3.*
google-cloud-storage==2.*
